Home

RvSpectML

RvSpectML

Modules

Exported Functions

Exported Types

Private Types & Functions

Index

Modules

RvSpectML.RvSpectML — Module.

RvSpectML.jl is a Julia package for measuring radial velocities (RVs) from stellar spectroscopic timeseries via machine learning (ML).

source

RvSpectML.CCF — Module.

Module for computing CCFs

source

RvSpectML.RVFromCCF — Module.

Module for estimating the radial velocity based on the CCF

source

RvSpectML.DCPCA — Module.

Module for performing a Doppler-constrained PCA analysis

For algorithm information, see:

Jones, Stenning, Ford et al. https://arxiv.org/abs/1711.01318

See also: Gilbertson, Ford, Jones & Stenning 2020 https://arxiv.org/abs/2009.01085

Author: Eric Ford Date: September 2020

source

Modules = [RvSpectML.LinearInterpolation, RvSpectML.SincInterpolation, RvSpectML.GPInterpolation, RvSpectML.TemporalGPInterpolation ]
Order = [:module]

Instrument-specific Modules

RvSpectML.EXPRES — Module.

Delegates loading functions & traits for the EXPRES spectrograph http://exoplanets.astro.yale.edu/expresBlog/ https://ui.adsabs.harvard.edu/abs/2016SPIE.9908E..6TJ/abstract Author: Eric Ford and collaborators Created: August 2020

source

RvSpectML.HARPSN — Module.

Delegates loading functions & traits for the HARPS-N spectrograph Author: Eric Ford Created: August 2020

source

Exported Functions

General purpose

Base.length — Method.

Return number of ChunkLists in a ChunkListTimeseries

source

Base.length — Method.

Return number of chunks in ChunkList

source

RvSpectML.absorption_line — Method.

absorption_line(x; mid, width, depth)

Return a Gaussian absorption line profile evaluated at x.

source

RvSpectML.apply_doppler_boost! — Function.

apply_doppler_boost!(spectrum, doppler_factor) -> typeof(spectrum)
apply_doppler_boost!(spectra, df) -> typeof(spectra)

Apply Doppler boost to spectra's λ's and update its metadata[:doppler_factor], so it will know how to undo the transform.

Arguments:

spectrum::AbstractSpectra: spectrum to be boosted

doppler_factor::Real: boost factor (1 = noop)

or

spectra::AbstractArray{<:AbstractSpectra}: spectra to be boosted

df::DataFrame: provides :drift and :ssb_rv (in m/s) for calculating the Doppler boost for each spectrum

TODO: Improve documentation formatting. This can serve as a template.

source

RvSpectML.bin_spectra_consecutive — Method.

binspectraconsecutive Bins consecutive spectra from a SpectralTimeSeriesCommonWavelengths object

WARNING: Simply takes consecutive spectra, so some bins may be from spectra that weren't taken close together. TODO: Create version that pays attention to timestamps.

source

RvSpectML.bin_spectra_max_Δt — Method.

binspectramax_Δt Bins spectra from a SpectralTimeSeriesCommonWavelengths object with a maximum spacing between observation times

source

RvSpectML.bin_spectra_nightly — Method.

binspectranightly Bins spectra from a SpectralTimeSeriesCommonWavelengths object with a maximum spacing between observation times

source

RvSpectML.bin_times_consecutive — Method.

bintimesconsecutive( times, n ) Computes mean times from conseuctive bins of n times (to go with binconsecutivespectra). Returns floor(length(times)/n) elements.

WARNING: Simply takes consecutive times, so some bins may be from spectra that weren't taken close together. TODO: Create version that pays attention to timestamps.

source

RvSpectML.calc_d2fluxdlnlambda2 — Method.

Estimate numerical second derivative of fluxes given wavelengths.

source

RvSpectML.calc_dfluxdlnlambda — Method.

Estimate numerical derivative of fluxes given wavelengths.

source

RvSpectML.calc_doppler_factor — Function.

calcdopplerfactor(rv; v_perp)

Return the Doppler boost factor (non-relativistic) for rv in km/s.

source

RvSpectML.calc_mean_dfluxdlnlambda — Method.

Return mean numerical derivative of fluxes based on a common set of wavelengths. Inputs: flux & var (2d) and λ (1d)

source

RvSpectML.calc_mean_spectrum — Method.

Return mean flux (averaging over observations at different times, variance weighted) based on a common set of wavelengths. Inputs: flux & var (2d: pixel, time)

source

RvSpectML.calc_normalization — Method.

Calc normalization of spectra based on average flux in a ChunkList.

source

RvSpectML.code_to_include_param_jl — Function.

codetoreadparamjl( pathto_search )

Returns a Code object. After res = code_toread_param_jl( path_to_search ), execute eval(res) to actually include the param.jl file. This is useful since it allows variables to be placed into caller's namespace.

Warning: Malicious users could insert arbitrary code into param.jl. Don't be a malicous user.

source

RvSpectML.filter_bad_chunks — Method.

Return (chunktimeseries, linelist) that have been trimmed of any chunks that are bad based on any spectra in the chunktimeseries. chunktimeseries: ChunkListTimeseries linelinst: DataFrame w/ lambdalo, lambda_hi verbose: print debugging info (false)

source

RvSpectML.find_overlapping_chunks — Method.

Return indices of any chunks in df that have overlapping lambdahi[i] and lambdalo[i+1].

source

RvSpectML.get_λ_range — Method.

Return the largest minimum wavelength and smallest maximum wavelength across an array of spectra. Calls getλrange(AbstractSpectra2D) that should be specialized for each instrument.

source

RvSpectML.make_chunk_list — Method.

Return a ChunkList of best regions of spectrum with lines in lineline. linelist is a DataFrame containing :lambdalo and :lambdahi. Pads edges by Δ.

source

RvSpectML.make_grid_for_chunk — Method.

makegridfor_chunk Create a range with equal spacing between points with end points set based on union of all chunks in timeseries.

Arguments:

timeseries: ChunkListTimeseries

chunk index:

oversample_factor: (1)

source

RvSpectML.make_manifest — Method.

makemanifest(targetsubdir::String, Inst::Module; [opts] ) Returns a dataframe containing a list of files to be read and some metadata (e.g., observation times)

Optional arguements

maxspectratouse (defaultmaxspectrato_use)

pathstosearch (defaultpathsto_search)

verbose = true

Warning: Malicious users could insert arbitrary code into data_paths.jl. Don't be a malicous user.

source

RvSpectML.make_orders_into_chunks — Function.

makeordersinto_chunks

Return a ChunkList with a region of spectrum from each order in orderstouse.

Arguments

spectra<:AbstractSpectra

inst: Instrument trait that provides default values

Optional arguments

orderstouse: Range or Array (orderstouse(inst))

pixelstouse: Array of Ranges (each from mincol to maxcol)

or

mincol: (mincol_default(inst,order)) and

maxcol: (maxcol_default(inst,order))

source

RvSpectML.make_vec_metadata_from_spectral_timeseries — Method.

Extract the metadata from a time series of spectra and return it as an array.

source

RvSpectML.merge_chunks — Method.

Return DataFrame with specifications for each chunk which will contain one or more lines. Input: linelist a DataFrame with: - lambdalo, lambdahi, lambda, depth Output: DataFrame with - lambdalo & lambdahi: boundaries for chunk - lambda & linedepths: arrays with info about each line

source

RvSpectML.normalize_spectrum! — Method.

Normalize spectrum, multiplying fluxes by scale_fac.

source

RvSpectML.num_chunks — Method.

Number of chunks in first chunklist in chunklist_timeseries.

source

RvSpectML.plot_spectrum_chunks — Method.

plotspectrumchunks( chunklist, chunks ) Returns a plot based of flux versus wavelength for selected chunks from the spectra in chunklist.

source

RvSpectML.read_header — Method.

Read header from FITS file and return Dict with contents.

source

RvSpectML.read_linelist_espresso — Method.

Read line list in ESPRESSO csv format.

ESPRESSO format: lambda and weight. Warning: ESPRESSO masks don't provide line depth and sometimes include one entry for a blend of lines.

source

RvSpectML.read_linelist_vald — Method.

Read line list in VALD csv format. VALD format: lambdalo, lambdaahi and depth.

source

RvSpectML.read_manifest — Method.

Read manifest containing filename, bjd, target, and optionally additional metadata from CSV file.

source

RvSpectML.read_mask_espresso — Method.

Read mask in ESPRESSO csv format.

ESPRESSO format: lambda and weight. Warning: ESPRESSO masks don't provide line depth and sometimes include one entry for a blend of lines.

source

RvSpectML.read_mask_vald — Method.

Read mask in VALD csv format. VALD format: lambdalo, lambdaahi and depth.

source

RvSpectML.λ_air_to_vac — Method.

Convert air wavelength (in Å) to vacuum wavelength Ref: https://www.astro.uu.se/valdwiki/Air-to-vacuum%20conversion VALD3 tools use the following solution derived by N. Piskunov

source

RvSpectML.λ_vac_to_air — Method.

Convert vacuum wavelength (in Å) to air wavelength Ref: Donald Morton (2000, ApJ. Suppl., 130, 403) via https://www.astro.uu.se/valdwiki/Air-to-vacuum%20conversion

source

RV-Related Algorithms

RvSpectML.CCF.calc_ccf_chunk — Method.

calcccfchunk( chunk, line_list::ALL ) Convenience function to compute CCF for one chunk of spectrum.

Inputs:

chunk

line_list

Optional Arguments:

mask_shape (TopHatCCFMask())

ccf_plan (BasicCCFPlan())

Return:

CCF for one chunk of spectrum, evaluated using mask_shape and plan

source

RvSpectML.CCF.calc_ccf_chunklist — Method.

calcccfchunk( chunklist, line_list ) Convenience function to compute CCF based on a spectrum's chunklist.

Inputs:

chunklist

line_list

Optional Arguments:

mask_shape (TopHatCCFMask())

ccf_plan (BasicCCFPlan())

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using mask_shape and plan.

source

RvSpectML.CCF.calc_ccf_chunklist_timeseries — Method.

calcccfchunk( chunklisttimeseries, linelist ) Convenience function to compute CCF for a timeseries of spectra, each with a chunklist. Uses multiple threads if avaliable.

Inputs:

chunklist_timeseries

line_list

Optional Arguments:

mask_shape (TopHatCCFMask())

ccf_plan (BasicCCFPlan())

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using mask_shape and plan.

source

RvSpectML.CCF.calc_ccf_v_grid — Method.

calcccfvgrid( plan ) Return range with 2n+1 points between -vmax and v_max where CCF is to be evaluated. Units based on those in plan.

source

RvSpectML.CCF.ccf_1D! — Method.

ccf_1D!(ccf_out, λs, fluxes, line_list; mask_shape, plan )

Compute the cross correlation function of a spectrum with a mask.

Inputs:

ccfout: 1-d array of size length(calcccfvgrid(plan)) to store output

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

line_linst: Each mask entry consists of the left and right end of tophats and a weight.

Optional Arguments:

mask_shape: shape of mask to use (TopHatCCFMask())

plan: parameters for computing ccf (BasicCCFPlan())

source

RvSpectML.CCF.ccf_1D — Method.

ccf_1D!(ccf_out, λs, fluxes, line_list; mask_shape, plan )

Compute the cross correlation function of a spectrum with a mask.

Inputs:

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

line_linst: Each mask entry consists of the left and right end of tophats and a weight.

Optional Arguments:

mask_shape: shape of mask to use (TopHatCCFMask())

plan: parameters for computing ccf (BasicCCFPlan())

Returns:

1-d array of size length(calcccfv_grid(plan))

source

RvSpectML.RVFromCCF.measure_rv_from_ccf — Method.

measure_rv_from_ccf(vels, ccf; fit_type="gaussian")

Fit a gaussian to the CCF to calculate a velocity using the specified method. Valid arguments for fit_type include "gaussian", "quadratic", and "centroid".

source

RvSpectML.RVFromCCF.measure_rvs_from_ccf — Method.

measure_rvs_from_ccf(vels, ccf; fit_type="gaussian")

At each time, compute the cross correlation function and fit to calculate a velocity using the specified method. Valid arguments for fit_type include "gaussian", "quadratic", and "centroid".

source

Interpolation Algorithms

Modules = [RvSpectML.LinearInterpolation, RvSpectML.SincInterpolation, RvSpectML.GPInterpolation, RvSpectML.TemporalGPInterpolation ]
Private = false
Order = [:function ]

Instrument-specific

RvSpectML.EXPRES.read_data — Function.

Read EXPRES data from FITS file, and return in a Spectra2DBasic object.

source

RvSpectML.NEID.make_manifest — Method.

Create Dataframe containing filenames and key data for all files neid*.fits in directory

source

RvSpectML.NEID.read_barycentric_corrections! — Function.

Read CSV of NEID barycentric corrections, interpolate to bjd's in df and insert into df[:,ssb_rv].

source

RvSpectML.NEID.read_data — Function.

Read NEID (non-solar) data from FITS file, and return in a Spectra2DBasic object.

source

RvSpectML.NEID.read_drift_corrections! — Function.

Read CSV of NEID drift corrections, interpolate to bjd's in df and insert into df[:,drift].

source

RvSpectML.NEID.read_metadata — Method.

Create Dict containing filename and default metadata from file.

source

RvSpectML.NEID.read_solar_data — Function.

Read NEID Solar data from FITS file, and return in a Spectra2DBasic object.

source

Exported Types

General purpose

RvSpectML.AbstractChuckOfSpectrum — Type.

Abstract type for any ChunkOfSpectrum

source

RvSpectML.AbstractInstrument — Type.

Abstract Base type for Instruments

source

RvSpectML.AbstractInstrument1D — Type.

Abstract Base type for 1-D extracted spectra Should be specialized for specific instruments.

source

RvSpectML.AbstractInstrument2D — Type.

Abstract Base type for 2-D extracted spectra. Should be specialized for each instrument

source

RvSpectML.AbstractSpectra — Type.

Abstract type for any Spectrum (or region of spectrum)

source

RvSpectML.AbstractSpectra1D — Type.

Abstract type for any 1-d Spectrum (or region of spectrum)

source

RvSpectML.AbstractSpectra2D — Type.

Abstract type for any 2-d Spectrum

source

RvSpectML.AbstractSpectralTimeSeriesCommonWavelengths — Type.

Abstract type for a time series of spectra that share a common wavelength grid.

source

RvSpectML.ChuckOfSpectrum — Type.

ChuckOfSpectrum for views into Spectra1DBasic or Spectra2DBasic

source

RvSpectML.ChunkList — Type.

Struct containing an array of ChuckOfSpectrum

source

RvSpectML.ChunkListTimeseries — Type.

Mtching lists of times and array of ChunkLists

source

RvSpectML.ChunkWidthFixedΔlnλ — Type.

Functor to return a constant scale factor for chunk widths (regardless of wavelength passed)

source

RvSpectML.Generic1D — Type.

Trait for genric 1D Extracted spectra

source

RvSpectML.Generic2D — Type.

Trait for generic 2D Extracted spectra

source

RvSpectML.Spectra1DBasic — Type.

Basic struct for Spectra1D (or region of specturm) Instruments can specialize their own if additional data is avaliable.

source

RvSpectML.Spectra2DBasic — Type.

Basic struct for Spectra2D (or region of specturm) Instruments can specialize their own if additional data is avaliable.

source

RvSpectML.SpectralTimeSeriesCommonWavelengths — Type.

Time series of spectra that share a common wavelength grid.

source

RvSpectML.CCF.AbstractCCFMaskShape — Type.

A struct implementing a specific mask shapes should be a subtype of AbstractCCFMaskShape.

source

RvSpectML.CCF.AbstractCCFPlan — Type.

A struct implementing a specific plans describing where the CCF is to be evaluated should be a subtype of AbstractCCFPlan.

source

RvSpectML.CCF.AbstractLineList — Type.

A struct implementing a line list should be a subtype of AbstractLineList.

source

RvSpectML.CCF.BasicCCFPlan — Type.

Basic plan for computing the CCF roughly between vcenter-vmax and vcenter+vmax with step size v_step.

source

RvSpectML.CCF.BasicCCFPlan — Method.

BasicCCFPlan

Optional arguments:

midpoint: (defaultvcenter)

step: (defaultvstep)

max: (defaultvmax)

source

RvSpectML.CCF.BasicLineList — Type.

A basic line list for passing to compute CCFs. Contains (views into) arrays specifying the minimum and maximum wavelength range and weight for each line.

source

RvSpectML.CCF.BasicLineList — Method.

BasicLineList( λ, weight )

source

RvSpectML.CCF.TopHatCCFMask — Type.

TopHatCCFMask The standard tophat mask with one parameter, it's half width as a velocity in m/s. Mask weights are stored separately in a line list.

source

RvSpectML.CCF.TopHatCCFMask — Type.

TopHatCCFMask( ; halfwidth=defaultv_width )

source

RvSpectML.CCF.TopHatCCFMask — Method.

TopHatCCFMask( inst )

source

RvSpectML.CCF.TopHatCCFMask — Method.

Functor for returning 1 for any Δv <= width.

source

Modules = [RvSpectML.LinearInterpolation, RvSpectML.SincInterpolation, RvSpectML.GPInterpolation, RvSpectML.TemporalGPInterpolation ]
Private = false
Order = [:type]

Instrument specific

RvSpectML.NEID.NEID1D — Type.

Trait for 1D Extracted spectra from NEID

source

RvSpectML.NEID.NEID2D — Type.

Trait for 2D Extracted spectra from NEID

source

RvSpectML.TheoreticalInstrument.TheoreticalInstrument1D — Type.

Trait for 1D spectra from a theoretical instrument

source

RvSpectML.TheoreticalInstrument.TheoreticalInstrument2D — Type.

Trait for a 2D spectra from a theoretical instrument

source

Private Types & Functions

General purpose

RvSpectML.ChunkWidthFuncOfλ — Type.

Functor to return a scale factor for chunk widths that depends on the chunk's central wavelength.

TODO: Replace with something useful, e.g., uses actual PSF behavior as function of wavelength

source

Base.getindex — Method.

Allow direct access to data, an AbstractArray of ChunkOfSpectrum's via [] operator

source

Base.setindex! — Method.

Allow direct access to data, an AbstractArray of ChunkOfSpectrum's via [] operator

source

RvSpectML.allequal — Method.

Return true if all elements of array are equal to each other.

source

RvSpectML.bin_times — Method.

bintimes() Bins times from a SpectralTimeSeriesCommonWavelengths object using the groupings from timeidx in the metadata.

source

RvSpectML.calc_d2fluxdlnlambda2! — Method.

Estimate numerical second derivative of fluxes given wavelengths.

source

RvSpectML.calc_dfluxdlnlambda! — Method.

Estimate numerical derivative of fluxes given wavelengths.

source

RvSpectML.calc_snr — Method.

Calculate total SNR in (region of) spectra.

source

RvSpectML.find_cols_to_fit — Method.

Return a range of columns indices with wavelengths between linelo and linehi

source

RvSpectML.find_cols_to_fit — Method.

Return a range of columns indices with wavelengths within Δ of line_center

source

RvSpectML.find_line_best — Function.

Return (pixels, order) pair that contain "best" region of spectra, based on highest SNR.

source

RvSpectML.find_orders_with_line — Method.

Return list of all orders that contain a pixel with wavelength lambda

source

RvSpectML.find_orders_with_line — Method.

Return list of all orders that contain all pixels with wavelengths between goallo and goalhi

source

RvSpectML.findall_line — Function.

Return list of (pixels, order) pairs that contain pixels with desireed wavelengths. Excludes locations that contain any pixels with var == NaN.

source

RvSpectML.findargminmax — Method.

findargminmax(a) Return (argmin, min, argmax, max) Adapapted from https://github.com/JuliaLang/julia/blob/697e782ab86bfcdd7fd15550241fe162c51d9f98/base/array.jl#L2191

source

RvSpectML.get_λs — Method.

" get_λs( grids::Vector{AbstractRange}, r::AbstractUnitRange ) Return range of wavelengths corresponding to indices in r from a vector of ranges. Useful for extracting wavelengths from a chunklist.

source

RvSpectML.interp_chunk_to_shifted_grid_linear! — Method.

interpchunktogridlinear!( fluxout, varout, chunkofspectrum, wavelengths ) Return spectra interpolated onto a grid of points using linear interpolation.

Arguments:

flux_out::AbstractArray

var_out::AbstractArray

chunk::AbstractChunkOfSpectrum

grid::AbstractRange or AbstractArray

Alters

flux_out

var_out::AbstractArray

Returns

flux_out

source

RvSpectML.interp_chunk_to_shifted_grid_sinc! — Method.

interpchunktogridsinc!( fluxout, varout, chunkofspectrum, wavelengths ) Return spectra interpolated onto a grid of points using sinc interpolation.

Arguments:

flux_out::AbstractArray

var_out::AbstractArray

chunk::AbstractChunkOfSpectrum

grid::AbstractRange or AbstractArray

Alters

flux_out

var_out::AbstractArray

Returns

flux_out

source

RvSpectML.make_template_spectra — Method.

maketemplatespectra( chunklisttimeseries, [ options ] ) Combine portions of spectra in a ChunkListTimeseries into a template stored as a SpectralTimeSeriesCommonWavelengths. Optionally remove radial velocities before combining spectra based on the :rv_est field in metadata.

Inputs:

chunklisttimeseries

Optional Parametes:

removervest (true)

oversample_factor (1)

smooth_factor (1)

alg (:TemporalGP)

Outputs

matrix: timeseries interpolated to each wavelength in the form of a SpectralTimeSeriesCommonWavelengths

f_mean: mean flux

var_mean: variance of mean flux (WARNING: Not computed accurately yet)

deriv: dflux/dlnλ averaged over spectra

deriv2: d²flux/dlnλ² averaged over spectra

source

RvSpectML.multiple_append! — Method.

A generalized version of the built in append!() function By Christian Gilbertson?

TODO: Ask Christian what the purpose of this is relative to std append

source

RvSpectML.normalize_spectra! — Method.

Normalize each spectrum based on sum of fluxes in chunk_timeseries region of each spectrum.

source

RvSpectML.predict_intrinsic_stellar_line_width — Method.

Estimate line width based on stellar Teff (K) and optionally v_rot (km/s). Output in km/s.

source

RvSpectML.read_metradata_from_fits — Function.

Read metadata in FITS header and return data for keys in fields_str/fields as a Dict.

source

RvSpectML.searchsortednearest — Function.

searchsortednearest(a<:AbstractVector, x::Real; assumesorted = false ) searchsortednearest(a<:AbstractVector, x<:AbstractVector; assumesorted = false )

Find the index of vector a where the value of a is closest to x. All vectors are assumed to already be sorted. To turn off assertions, set assume_sorted to true.

Credit: traktofon @ https://discourse.julialang.org/t/findnearest-function/4143/4 Vector Vector version by Christian Gilbertson? issorted assertion and optional assume_sorted added by Eric Ford

source

Base.length — Method.

Return length of line linst.

source

RvSpectML.CCF.calc_doppler_factor — Method.

calc_doppler_factor(vel)

Compute the longitudinal relativistic doppler factor given a velocity in meters per second.

source

RvSpectML.CCF.calc_order_ccf_chunklist_timeseries — Method.

calcccfchunk( chunklisttimeseries, linelist ) Convenience function to compute separate CCFs for each chunk of each spectrum in a timeseries. CCF is evaluated using mask_shape and plan. Uses multiple threads if avaliable.

Inputs:

chunklist_timeseries

line_list

Optional Arguments:

mask_shape (TopHatCCFMask())

ccf_plan (BasicCCFPlan())

Return:

A 3-d array containing the CCF at each (velocity, chunk, spectrum)

source

RvSpectML.CCF.calc_order_ccfs_chunklist — Method.

calcccfchunk( chunklisttimeseries, linelist ) Convenience function to compute separate CCFs for each chunk in a spectrum. CCF evaluated using mask_shape and plan.

Inputs:

chunklist_timeseries

line_list

Optional Arguments:

mask_shape (TopHatCCFMask())

ccf_plan (BasicCCFPlan())

Return:

A 2-d array containing the CCF at each (velocity, chunk)

source

RvSpectML.CCF.find_bin_edges_opt — Method.

findbinedges( pixel_centers )

Internal function used by project_mask!. TODO: OPT may be able to eliminate need for this memory allocation. Or at least preallocate it.

source

RvSpectML.CCF.project_mask_opt! — Method.

project_mask!( output, λs, line_list; shift_factor, mask_shape )

Compute the projection of the mask onto the 1D array of wavelengths (λs) at a given shift factor (default: 1). The mask is computed from a linelist and maskshape (default: tophat).

TODO: Implement/test other mask_shapes.

source

RvSpectML.RVFromCCF.measure_rv — Method.

measure_rv(wavs, spec, mask; fit_type="gaussian")

TODO: Need to update for CCF refactoring.

Compute the cross correlation function and fit to calculate a velocity using the specified method. Valid arguments for fit_type include "gaussian", "quadratic", and "centroid".

source

RvSpectML.RVFromCCF.prelim_fit — Method.

source

RvSpectML.RVFromCCF.rv_from_ccf_bestfit — Method.

Report RV based on minimum of CCF WARNING: This is highly pixelated and should be used for special purposes (e.g., initial guesses).

source

RvSpectML.RVFromCCF.rv_from_ccf_centroid — Method.

Report RV based on centroid near minimum of CCF. WARNING: Something's fishy with these.

source

RvSpectML.RVFromCCF.rv_from_ccf_gaussian — Method.

Report RV base don fit Gaussian near minimum of CCF.

source

RvSpectML.RVFromCCF.rv_from_ccf_quadratic — Method.

Report RV based on fitting quadratic near minimum of CCF.

source

Modules = [RvSpectML.LinearInterpolation, RvSpectML.SincInterpolation, RvSpectML.GPInterpolation, RvSpectML.TemporalGPInterpolation ]
Public = false
Order = [:type, :function]

Instrument specific

RvSpectML.EXPRES.make_manifest — Method.

Create Dataframe containing filenames and key data for all files neid*.fits in directory

source

RvSpectML.EXPRES.read_metadata — Method.

Create Dict containing filename and default metadata from file.

source

RvSpectML.NEID.read_differential_extinctions! — Function.

Read space delimited file with differential extinction corrections, interpolate to bjd's in df and insert into df[:,diffextrv].

source

Index

RvSpectML.CCF

RvSpectML.DCPCA

RvSpectML.EXPRES

RvSpectML.HARPSN

RvSpectML.RVFromCCF

RvSpectML.RvSpectML

RvSpectML.AbstractChuckOfSpectrum

RvSpectML.AbstractInstrument

RvSpectML.AbstractInstrument1D

RvSpectML.AbstractInstrument2D

RvSpectML.AbstractSpectra

RvSpectML.AbstractSpectra1D

RvSpectML.AbstractSpectra2D

RvSpectML.AbstractSpectralTimeSeriesCommonWavelengths

RvSpectML.CCF.AbstractCCFMaskShape

RvSpectML.CCF.AbstractCCFPlan

RvSpectML.CCF.AbstractLineList

RvSpectML.CCF.BasicCCFPlan

RvSpectML.CCF.BasicCCFPlan

RvSpectML.CCF.BasicLineList

RvSpectML.CCF.BasicLineList

RvSpectML.CCF.TopHatCCFMask

RvSpectML.CCF.TopHatCCFMask

RvSpectML.CCF.TopHatCCFMask

RvSpectML.CCF.TopHatCCFMask

RvSpectML.ChuckOfSpectrum

RvSpectML.ChunkList

RvSpectML.ChunkListTimeseries

RvSpectML.ChunkWidthFixedΔlnλ

RvSpectML.ChunkWidthFuncOfλ

RvSpectML.Generic1D

RvSpectML.Generic2D

RvSpectML.NEID.NEID1D

RvSpectML.NEID.NEID2D

RvSpectML.Spectra1DBasic

RvSpectML.Spectra2DBasic

RvSpectML.SpectralTimeSeriesCommonWavelengths

RvSpectML.TheoreticalInstrument.TheoreticalInstrument1D

RvSpectML.TheoreticalInstrument.TheoreticalInstrument2D

Base.getindex

Base.length

Base.length

Base.length

Base.setindex!

RvSpectML.CCF.calc_ccf_chunk

RvSpectML.CCF.calc_ccf_chunklist

RvSpectML.CCF.calc_ccf_chunklist_timeseries

RvSpectML.CCF.calc_ccf_v_grid

RvSpectML.CCF.calc_doppler_factor

RvSpectML.CCF.calc_order_ccf_chunklist_timeseries

RvSpectML.CCF.calc_order_ccfs_chunklist

RvSpectML.CCF.ccf_1D

RvSpectML.CCF.ccf_1D!

RvSpectML.CCF.find_bin_edges_opt

RvSpectML.CCF.project_mask_opt!

RvSpectML.EXPRES.make_manifest

RvSpectML.EXPRES.read_data

RvSpectML.EXPRES.read_metadata

RvSpectML.NEID.make_manifest

RvSpectML.NEID.read_barycentric_corrections!

RvSpectML.NEID.read_data

RvSpectML.NEID.read_differential_extinctions!

RvSpectML.NEID.read_drift_corrections!

RvSpectML.NEID.read_metadata

RvSpectML.NEID.read_solar_data

RvSpectML.RVFromCCF.measure_rv

RvSpectML.RVFromCCF.measure_rv_from_ccf

RvSpectML.RVFromCCF.measure_rvs_from_ccf

RvSpectML.RVFromCCF.prelim_fit

RvSpectML.RVFromCCF.rv_from_ccf_bestfit

RvSpectML.RVFromCCF.rv_from_ccf_centroid

RvSpectML.RVFromCCF.rv_from_ccf_gaussian

RvSpectML.RVFromCCF.rv_from_ccf_quadratic

RvSpectML.absorption_line

RvSpectML.allequal

RvSpectML.apply_doppler_boost!

RvSpectML.bin_spectra_consecutive

RvSpectML.bin_spectra_max_Δt

RvSpectML.bin_spectra_nightly

RvSpectML.bin_times

RvSpectML.bin_times_consecutive

RvSpectML.calc_d2fluxdlnlambda2

RvSpectML.calc_d2fluxdlnlambda2!

RvSpectML.calc_dfluxdlnlambda

RvSpectML.calc_dfluxdlnlambda!

RvSpectML.calc_doppler_factor

RvSpectML.calc_mean_dfluxdlnlambda

RvSpectML.calc_mean_spectrum

RvSpectML.calc_normalization

RvSpectML.calc_snr

RvSpectML.code_to_include_param_jl

RvSpectML.filter_bad_chunks

RvSpectML.find_cols_to_fit

RvSpectML.find_cols_to_fit

RvSpectML.find_line_best

RvSpectML.find_orders_with_line

RvSpectML.find_orders_with_line

RvSpectML.find_overlapping_chunks

RvSpectML.findall_line

RvSpectML.findargminmax

RvSpectML.get_λ_range

RvSpectML.get_λs

RvSpectML.interp_chunk_to_shifted_grid_linear!

RvSpectML.interp_chunk_to_shifted_grid_sinc!

RvSpectML.make_chunk_list

RvSpectML.make_grid_for_chunk

RvSpectML.make_manifest

RvSpectML.make_orders_into_chunks

RvSpectML.make_template_spectra

RvSpectML.make_vec_metadata_from_spectral_timeseries

RvSpectML.merge_chunks

RvSpectML.multiple_append!

RvSpectML.normalize_spectra!

RvSpectML.normalize_spectrum!

RvSpectML.num_chunks

RvSpectML.plot_spectrum_chunks

RvSpectML.predict_intrinsic_stellar_line_width

RvSpectML.read_header

RvSpectML.read_linelist_espresso

RvSpectML.read_linelist_vald

RvSpectML.read_manifest

RvSpectML.read_mask_espresso

RvSpectML.read_mask_vald

RvSpectML.read_metradata_from_fits

RvSpectML.searchsortednearest

RvSpectML.λ_air_to_vac

RvSpectML.λ_vac_to_air