Modules

RvSpectML Modules

RvSpectML Modules

RV-Related Algorithms

Interpolation Algorithms

Instrument-specific Modules

Other Modules

RV-Related Algorithms

RvSpectML.CCF — Module.

Module for computing CCFs

source

RvSpectML.RVFromCCF — Module.

Module for estimating the radial velocity based on the CCF

source

RvSpectML.DCPCA — Module.

Module for performing a Doppler-constrained PCA analysis

For algorithm information, see:

Jones, Stenning, Ford et al. https://arxiv.org/abs/1711.01318

See also: Gilbertson, Ford, Jones & Stenning 2020 https://arxiv.org/abs/2009.01085

Author: Eric Ford Date: September 2020

source

RvSpectML.Scalpels — Module.

Module for performing Scalpels (Self-Correlation Analysis of Line Profiles for Extracting Low-amplitude Shifts) based on a CCF timeseries.

For algorithm information, see Collier-Cameron, Ford, Shahaf et al. 2020

Author: Eric Ford Date: September 2020

source

Interpolation Algorithms

RvSpectML.LinearInterpolation — Module.

Module with wrappers for performing linear interpolation..

Author: Eric Ford

source

RvSpectML.SincInterpolation — Module.

Module for performing Sinc interpolation From https://github.com/christiangil/GP-Research/blob/master/julia/src/interpolationfunctions.jl https://github.com/christiangil/GP-Research/blob/master/julia/src/generalfunctions.jl

Authors: Joe Ninan (original) Christian Gilbertson (Converted to Julia and optimized) Eric Ford (Further adapted/optimizations)

source

RvSpectML.TemporalGPInterpolation — Module.

Module for interpolating via Gaussian Process Regression based on Stheno and TemporalGPs packages.

source

Instrument-specific Modules

RvSpectML.EXPRES — Module.

Module providing types and traits and customized functions for the EXPRES Spectrograph. http://exoplanets.astro.yale.edu/expresBlog/ https://ui.adsabs.harvard.edu/abs/2016SPIE.9908E..6TJ/abstract

source

RvSpectML.HARPSN — Module.

Module providing types and traits and customized functions for the HARPS-N Spectrograph.

source

RvSpectML.NEID — Module.

Module providing types and traits and customized functions for the NEID Spectrograph.

source

RvSpectML.TheoreticalInstrument — Module.

Module providing types and traits and customized functions for a generic theoretical spectrograph.

source

Other Modules

RvSpectML.Pipeline — Module.

The PipelinePlan stores what work needs to be done, what plots should be made, and what data/plots should be saved to disk. Users will querty it via needto(plan, symbol), makeplot(plan, symbol), saveplot(plan, symbol), and savedata(plan, symbol). Users can update the plan with makeplot!(plan, symbol), dontmake_plot!(plan, symbol), etc.

In principle, we could cache data here, too. But I'm not sure if we'll actulaly use that. So consider the cache as experimental.

source