Types

Types Exported by RvSpectML

Types Exported by RvSpectML

General purpose

RV-Related Algorithms

Interpolation Algorithms

Instrument specific

Other

General purpose

RvSpectML.AbstractChuckOfSpectrum — Type.

Abstract type for any ChunkOfSpectrum

source

RvSpectML.AbstractInstrument — Type.

Abstract Base type for Instruments

source

RvSpectML.AbstractInstrument1D — Type.

Abstract Base type for 1-D extracted spectra Should be specialized for specific instruments.

source

RvSpectML.AbstractInstrument2D — Type.

Abstract Base type for 2-D extracted spectra. Should be specialized for each instrument

source

RvSpectML.AbstractSpectra — Type.

Abstract type for any Spectrum (or region of spectrum)

source

RvSpectML.AbstractSpectra1D — Type.

Abstract type for any 1-d Spectrum (or region of spectrum)

source

RvSpectML.AbstractSpectra2D — Type.

Abstract type for any 2-d Spectrum

source

RvSpectML.AbstractSpectralTimeSeriesCommonWavelengths — Type.

Abstract type for a time series of spectra that share a common wavelength grid.

source

RvSpectML.ChuckOfSpectrum — Type.

ChuckOfSpectrum for views into Spectra1DBasic or Spectra2DBasic

source

RvSpectML.ChunkList — Type.

Struct containing an array of ChuckOfSpectrum

source

RvSpectML.ChunkListTimeseries — Type.

Mtching lists of times and array of ChunkLists

source

RvSpectML.ChunkWidthFixedΔlnλ — Type.

Functor to return a constant scale factor for chunk widths (regardless of wavelength passed)

source

RvSpectML.Generic1D — Type.

Trait for genric 1D Extracted spectra

source

RvSpectML.Generic2D — Type.

Trait for generic 2D Extracted spectra

source

RvSpectML.Spectra1DBasic — Type.

Basic struct for Spectra1D (or region of specturm) Instruments can specialize their own if additional data is avaliable.

source

RvSpectML.Spectra2DBasic — Type.

Basic struct for Spectra2D (or region of specturm) Instruments can specialize their own if additional data is avaliable.

source

RvSpectML.SpectralTimeSeriesCommonWavelengths — Type.

Time series of spectra that share a common wavelength grid.

source

RV-Related Algorithms

RvSpectML.CCF.AbstractCCFMaskShape — Type.

A struct implementing a specific mask shapes should be a subtype of AbstractCCFMaskShape.

source

RvSpectML.CCF.AbstractCCFPlan — Type.

A struct implementing a specific plans describing where the CCF is to be evaluated should be a subtype of AbstractCCFPlan.

source

RvSpectML.CCF.AbstractLineList — Type.

A struct implementing a line list should be a subtype of AbstractLineList.

source

RvSpectML.CCF.BasicCCFPlan — Type.

Basic plan for computing the CCF roughly between vcenter-vmax and vcenter+vmax with step size v_step.

source

RvSpectML.CCF.BasicCCFPlan — Method.

BasicCCFPlan

Optional arguments:

midpoint: (defaultvcenter)

step: (defaultvstep)

max: (defaultvmax)

source

RvSpectML.CCF.BasicLineList — Type.

A basic line list for passing to compute CCFs. Contains (views into) arrays specifying the minimum and maximum wavelength range and weight for each line.

source

RvSpectML.CCF.BasicLineList — Method.

BasicLineList( λ, weight )

source

RvSpectML.CCF.TopHatCCFMask — Type.

TopHatCCFMask The standard tophat mask with one parameter, it's half width as a velocity in m/s. Mask weights are stored separately in a line list.

source

RvSpectML.CCF.TopHatCCFMask — Type.

TopHatCCFMask( ; halfwidth=defaultv_width )

source

RvSpectML.CCF.TopHatCCFMask — Method.

TopHatCCFMask( inst )

source

RvSpectML.CCF.TopHatCCFMask — Method.

Functor for returning 1 for any Δv <= width.

source

Interpolation Algorithms

Instrument specific

RvSpectML.NEID.NEID1D — Type.

Trait for 1D Extracted spectra from NEID

source

RvSpectML.NEID.NEID2D — Type.

Trait for 2D Extracted spectra from NEID

source

RvSpectML.TheoreticalInstrument.TheoreticalInstrument1D — Type.

Trait for 1D spectra from a theoretical instrument

source

RvSpectML.TheoreticalInstrument.TheoreticalInstrument2D — Type.

Trait for a 2D spectra from a theoretical instrument

source

Other