Home

RvSpectML

RvSpectML

Modules

Exported Functions

Exported Types

RvSpectML

Internals Types & Functions

General purpose

Radial Velocity Related

Interpolation

Instrument specific

Index

Modules

RvSpectML.RvSpectML — Module.

RvSpectML.jl is a Julia package for measuring radial velocities (RVs) from stellar spectroscopic timeseries via machine learning (ML).

source

RV-Related Algorithms

RvSpectML.CCF — Module.

Module for computing CCFs

source

RvSpectML.RVFromCCF — Module.

Module for estimating the radial velocity based on the CCF

source

RvSpectML.DCPCA — Module.

Module for performing a Doppler-constrained PCA analysis

For algorithm information, see:

Jones, Stenning, Ford et al. https://arxiv.org/abs/1711.01318

See also: Gilbertson, Ford, Jones & Stenning 2020 https://arxiv.org/abs/2009.01085

Author: Eric Ford Date: September 2020

source

Interpolation Algorithms

RvSpectML.SincInterpolation — Module.

Module for performing Sinc interpolation From https://github.com/christiangil/GP-Research/blob/master/julia/src/interpolationfunctions.jl https://github.com/christiangil/GP-Research/blob/master/julia/src/generalfunctions.jl

Authors: Joe Ninan (original) Christian Gilbertson (Converted to Julia and optimized) Eric Ford (Further adapted/optimizations)

source

RvSpectML.TemporalGPInterpolation — Module.

Module for interpolating via Gaussian Process Regression based on Stheno and TemporalGPs packages.

source

Instrument-specific Modules

RvSpectML.EXPRES — Module.

Delegates loading functions & traits for the EXPRES spectrograph http://exoplanets.astro.yale.edu/expresBlog/ https://ui.adsabs.harvard.edu/abs/2016SPIE.9908E..6TJ/abstract Author: Eric Ford and collaborators Created: August 2020

source

RvSpectML.HARPSN — Module.

Delegates loading functions & traits for the HARPS-N spectrograph Author: Eric Ford Created: August 2020

source

Exported Functions

General purpose

Base.length — Method.

Return number of ChunkLists in a ChunkListTimeseries

source

Base.length — Method.

Return number of chunks in ChunkList

source

RvSpectML.absorption_line — Method.

absorption_line(x; mid, width, depth)

Return a Gaussian absorption line profile evaluated at x.

source

RvSpectML.apply_doppler_boost! — Function.

apply_doppler_boost!(spectrum, doppler_factor) -> typeof(spectrum)
apply_doppler_boost!(spectra, df) -> typeof(spectra)

Apply Doppler boost to spectra's λ's and update its metadata[:doppler_factor], so it will know how to undo the transform.

Arguments:

spectrum::AbstractSpectra: spectrum to be boosted

doppler_factor::Real: boost factor (1 = noop)

or

spectra::AbstractArray{<:AbstractSpectra}: spectra to be boosted

df::DataFrame: provides :drift and :ssb_rv (in m/s) for calculating the Doppler boost for each spectrum

TODO: Improve documentation formatting. This can serve as a template.

source

RvSpectML.bin_spectra_consecutive — Method.

binspectraconsecutive Bins consecutive spectra from a SpectralTimeSeriesCommonWavelengths object

WARNING: Simply takes consecutive spectra, so some bins may be from spectra that weren't taken close together. TODO: Create version that pays attention to timestamps.

source

RvSpectML.bin_spectra_max_Δt — Method.

binspectramax_Δt Bins spectra from a SpectralTimeSeriesCommonWavelengths object with a maximum spacing between observation times

source

RvSpectML.bin_spectra_nightly — Method.

binspectranightly Bins spectra from a SpectralTimeSeriesCommonWavelengths object with a maximum spacing between observation times

source

RvSpectML.bin_times_consecutive — Method.

bintimesconsecutive( times, n ) Computes mean times from conseuctive bins of n times (to go with binconsecutivespectra). Returns floor(length(times)/n) elements.

WARNING: Simply takes consecutive times, so some bins may be from spectra that weren't taken close together. TODO: Create version that pays attention to timestamps.

source

RvSpectML.calc_d2fluxdlnlambda2 — Method.

Estimate numerical second derivative of fluxes given wavelengths.

source

RvSpectML.calc_dfluxdlnlambda — Method.

Estimate numerical derivative of fluxes given wavelengths.

source

RvSpectML.calc_doppler_factor — Function.

calcdopplerfactor(rv; v_perp)

Return the Doppler boost factor (non-relativistic) for rv in km/s.

source

RvSpectML.calc_mean_dfluxdlnlambda — Method.

Return mean numerical derivative of fluxes based on a common set of wavelengths. Inputs: flux & var (2d) and λ (1d)

source

RvSpectML.calc_mean_spectrum — Method.

Return mean flux (averaging over observations at different times, variance weighted) based on a common set of wavelengths. Inputs: flux & var (2d: pixel, time)

source

RvSpectML.calc_normalization — Method.

Calc normalization of spectra based on average flux in a ChunkList.

source

RvSpectML.code_to_include_param_jl — Function.

codetoreadparamjl( pathto_search )

Returns a Code object. After res = code_toread_param_jl( path_to_search ), execute eval(res) to actually include the param.jl file. This is useful since it allows variables to be placed into caller's namespace.

Warning: Malicious users could insert arbitrary code into param.jl. Don't be a malicous user.

source

RvSpectML.filter_bad_chunks — Method.

Return (chunktimeseries, linelist) that have been trimmed of any chunks that are bad based on any spectra in the chunktimeseries. chunktimeseries: ChunkListTimeseries linelinst: DataFrame w/ lambdalo, lambda_hi verbose: print debugging info (false)

source

RvSpectML.find_overlapping_chunks — Method.

Return indices of any chunks in df that have overlapping lambdahi[i] and lambdalo[i+1].

source

RvSpectML.get_λ_range — Method.

Return the largest minimum wavelength and smallest maximum wavelength across an array of spectra. Calls getλrange(AbstractSpectra2D) that should be specialized for each instrument.

source

RvSpectML.make_chunk_list — Method.

Return a ChunkList of best regions of spectrum with lines in lineline. linelist is a DataFrame containing :lambdalo and :lambdahi. Pads edges by Δ.

source

RvSpectML.make_grid_for_chunk — Method.

makegridfor_chunk Create a range with equal spacing between points with end points set based on union of all chunks in timeseries.

Arguments:

timeseries: ChunkListTimeseries

chunk index:

oversample_factor: (1)

source

RvSpectML.make_manifest — Method.

makemanifest(targetsubdir::String, Inst::Module; [opts] ) Returns a dataframe containing a list of files to be read and some metadata (e.g., observation times)

Optional arguements

maxspectratouse (defaultmaxspectrato_use)

pathstosearch (defaultpathsto_search)

verbose = true

Warning: Malicious users could insert arbitrary code into data_paths.jl. Don't be a malicous user.

source

RvSpectML.make_orders_into_chunks — Function.

makeordersinto_chunks

Return a ChunkList with a region of spectrum from each order in orderstouse.

Arguments

spectra<:AbstractSpectra

inst: Instrument trait that provides default values

Optional arguments

orderstouse: Range or Array (orderstouse(inst))

pixelstouse: Array of Ranges (each from mincol to maxcol)

or

mincol: (mincol_default(inst,order)) and

maxcol: (maxcol_default(inst,order))

source

RvSpectML.make_vec_metadata_from_spectral_timeseries — Method.

Extract the metadata from a time series of spectra and return it as an array.

source

RvSpectML.merge_chunks — Method.

Return DataFrame with specifications for each chunk which will contain one or more lines. Input: linelist a DataFrame with: - lambdalo, lambdahi, lambda, depth Output: DataFrame with - lambdalo & lambdahi: boundaries for chunk - lambda & linedepths: arrays with info about each line

source

RvSpectML.normalize_spectrum! — Method.

Normalize spectrum, multiplying fluxes by scale_fac.

source

RvSpectML.num_chunks — Method.

Number of chunks in first chunklist in chunklist_timeseries.

source

RvSpectML.plot_spectrum_chunks — Method.

plotspectrumchunks( chunklist, chunks ) Returns a plot based of flux versus wavelength for selected chunks from the spectra in chunklist.

source

RvSpectML.read_header — Method.

Read header from FITS file and return Dict with contents.

source

RvSpectML.read_linelist_espresso — Method.

Read line list in ESPRESSO csv format.

ESPRESSO format: lambda and weight. Warning: ESPRESSO masks don't provide line depth and sometimes include one entry for a blend of lines.

source

RvSpectML.read_linelist_vald — Method.

Read line list in VALD csv format. VALD format: lambdalo, lambdaahi and depth.

source

RvSpectML.read_manifest — Method.

Read manifest containing filename, bjd, target, and optionally additional metadata from CSV file.

source

RvSpectML.read_mask_espresso — Method.

Read mask in ESPRESSO csv format.

ESPRESSO format: lambda and weight. Warning: ESPRESSO masks don't provide line depth and sometimes include one entry for a blend of lines.

source

RvSpectML.read_mask_vald — Method.

Read mask in VALD csv format. VALD format: lambdalo, lambdaahi and depth.

source

RvSpectML.λ_air_to_vac — Method.

Convert air wavelength (in Å) to vacuum wavelength Ref: https://www.astro.uu.se/valdwiki/Air-to-vacuum%20conversion VALD3 tools use the following solution derived by N. Piskunov

source

RvSpectML.λ_vac_to_air — Method.

Convert vacuum wavelength (in Å) to air wavelength Ref: Donald Morton (2000, ApJ. Suppl., 130, 403) via https://www.astro.uu.se/valdwiki/Air-to-vacuum%20conversion

source

RV-Related Algorithms

RvSpectML.CCF.calc_ccf_chunk — Method.

calcccfchunk( chunk, line_list::ALL ) Convenience function to compute CCF for one chunk of spectrum.

Inputs:

chunk

line_list

Optional Arguments:

mask_shape (TopHatCCFMask())

ccf_plan (BasicCCFPlan())

Return:

CCF for one chunk of spectrum, evaluated using mask_shape and plan

source

RvSpectML.CCF.calc_ccf_chunklist — Method.

calcccfchunk( chunklist, line_list ) Convenience function to compute CCF based on a spectrum's chunklist.

Inputs:

chunklist

line_list

Optional Arguments:

mask_shape (TopHatCCFMask())

ccf_plan (BasicCCFPlan())

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using mask_shape and plan.

source

RvSpectML.CCF.calc_ccf_chunklist_timeseries — Method.

calcccfchunk( chunklisttimeseries, linelist ) Convenience function to compute CCF for a timeseries of spectra, each with a chunklist. Uses multiple threads if avaliable.

Inputs:

chunklist_timeseries

line_list

Optional Arguments:

mask_shape (TopHatCCFMask())

ccf_plan (BasicCCFPlan())

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using mask_shape and plan.

source

RvSpectML.CCF.calc_ccf_v_grid — Method.

calcccfvgrid( plan ) Return range with 2n+1 points between -vmax and v_max where CCF is to be evaluated. Units based on those in plan.

source

RvSpectML.CCF.ccf_1D! — Method.

ccf_1D!(ccf_out, λs, fluxes, line_list; mask_shape, plan )

Compute the cross correlation function of a spectrum with a mask.

Inputs:

ccfout: 1-d array of size length(calcccfvgrid(plan)) to store output

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

line_linst: Each mask entry consists of the left and right end of tophats and a weight.

Optional Arguments:

mask_shape: shape of mask to use (TopHatCCFMask())

plan: parameters for computing ccf (BasicCCFPlan())

source

RvSpectML.CCF.ccf_1D — Method.

ccf_1D!(ccf_out, λs, fluxes, line_list; mask_shape, plan )

Compute the cross correlation function of a spectrum with a mask.

Inputs:

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

line_linst: Each mask entry consists of the left and right end of tophats and a weight.

Optional Arguments:

mask_shape: shape of mask to use (TopHatCCFMask())

plan: parameters for computing ccf (BasicCCFPlan())

Returns:

1-d array of size length(calcccfv_grid(plan))

source

RvSpectML.RVFromCCF.measure_rv_from_ccf — Method.

measure_rv_from_ccf(vels, ccf; fit_type="gaussian")

Fit a gaussian to the CCF to calculate a velocity using the specified method. Valid arguments for fit_type include "gaussian", "quadratic", and "centroid".

source

RvSpectML.RVFromCCF.measure_rvs_from_ccf — Method.

measure_rvs_from_ccf(vels, ccf; fit_type="gaussian")

At each time, compute the cross correlation function and fit to calculate a velocity using the specified method. Valid arguments for fit_type include "gaussian", "quadratic", and "centroid".

source

Interpolation Algorithms

RvSpectML.LinearInterpolation.make_interpolator_linear_var — Method.

Return interpolator for variances in spectra.

source

RvSpectML.SincInterpolation.spectra_interpolate — Method.

Original author: Joe Ninan Converted to Julia and optimized by Christian Gilbertson Further adapted/optimized by Eric Ford Additional optimizations possible by preallocating arrasy for minargs, Nminargs, minvalues, FilterValues and OldYCoords.

source

Instrument-specific

RvSpectML.EXPRES.read_data — Function.

Read EXPRES data from FITS file, and return in a Spectra2DBasic object.

source

RvSpectML.NEID.make_manifest — Method.

Create Dataframe containing filenames and key data for all files neid*.fits in directory

source

RvSpectML.NEID.read_barycentric_corrections! — Function.

Read CSV of NEID barycentric corrections, interpolate to bjd's in df and insert into df[:,ssb_rv].

source

RvSpectML.NEID.read_data — Function.

Read NEID (non-solar) data from FITS file, and return in a Spectra2DBasic object.

source

RvSpectML.NEID.read_drift_corrections! — Function.

Read CSV of NEID drift corrections, interpolate to bjd's in df and insert into df[:,drift].

source

RvSpectML.NEID.read_metadata — Method.

Create Dict containing filename and default metadata from file.

source

RvSpectML.NEID.read_solar_data — Function.

Read NEID Solar data from FITS file, and return in a Spectra2DBasic object.

source

Exported Types

General purpose

RvSpectML.AbstractChuckOfSpectrum — Type.

Abstract type for any ChunkOfSpectrum

source

RvSpectML.AbstractInstrument — Type.

Abstract Base type for Instruments

source

RvSpectML.AbstractInstrument1D — Type.

Abstract Base type for 1-D extracted spectra Should be specialized for specific instruments.

source

RvSpectML.AbstractInstrument2D — Type.

Abstract Base type for 2-D extracted spectra. Should be specialized for each instrument

source

RvSpectML.AbstractSpectra — Type.

Abstract type for any Spectrum (or region of spectrum)

source

RvSpectML.AbstractSpectra1D — Type.

Abstract type for any 1-d Spectrum (or region of spectrum)

source

RvSpectML.AbstractSpectra2D — Type.

Abstract type for any 2-d Spectrum

source

RvSpectML.AbstractSpectralTimeSeriesCommonWavelengths — Type.

Abstract type for a time series of spectra that share a common wavelength grid.

source

RvSpectML.ChuckOfSpectrum — Type.

ChuckOfSpectrum for views into Spectra1DBasic or Spectra2DBasic

source

RvSpectML.ChunkList — Type.

Struct containing an array of ChuckOfSpectrum

source

RvSpectML.ChunkListTimeseries — Type.

Mtching lists of times and array of ChunkLists

source

RvSpectML.ChunkWidthFixedΔlnλ — Type.

Functor to return a constant scale factor for chunk widths (regardless of wavelength passed)

source

RvSpectML.Generic1D — Type.

Trait for genric 1D Extracted spectra

source

RvSpectML.Generic2D — Type.

Trait for generic 2D Extracted spectra

source

RvSpectML.Spectra1DBasic — Type.

Basic struct for Spectra1D (or region of specturm) Instruments can specialize their own if additional data is avaliable.

source

RvSpectML.Spectra2DBasic — Type.

Basic struct for Spectra2D (or region of specturm) Instruments can specialize their own if additional data is avaliable.

source

RvSpectML.SpectralTimeSeriesCommonWavelengths — Type.

Time series of spectra that share a common wavelength grid.

source

RvSpectML.CCF.AbstractCCFMaskShape — Type.

A struct implementing a specific mask shapes should be a subtype of AbstractCCFMaskShape.

source

RvSpectML.CCF.AbstractCCFPlan — Type.

A struct implementing a specific plans describing where the CCF is to be evaluated should be a subtype of AbstractCCFPlan.

source

RvSpectML.CCF.AbstractLineList — Type.

A struct implementing a line list should be a subtype of AbstractLineList.

source

RvSpectML.CCF.BasicCCFPlan — Type.

Basic plan for computing the CCF roughly between vcenter-vmax and vcenter+vmax with step size v_step.

source

RvSpectML.CCF.BasicCCFPlan — Method.

BasicCCFPlan

Optional arguments:

midpoint: (defaultvcenter)

step: (defaultvstep)

max: (defaultvmax)

source

RvSpectML.CCF.BasicLineList — Type.

A basic line list for passing to compute CCFs. Contains (views into) arrays specifying the minimum and maximum wavelength range and weight for each line.

source

RvSpectML.CCF.BasicLineList — Method.

BasicLineList( λ, weight )

source

RvSpectML.CCF.TopHatCCFMask — Type.

TopHatCCFMask The standard tophat mask with one parameter, it's half width as a velocity in m/s. Mask weights are stored separately in a line list.

source

RvSpectML.CCF.TopHatCCFMask — Type.

TopHatCCFMask( ; halfwidth=defaultv_width )

source

RvSpectML.CCF.TopHatCCFMask — Method.

TopHatCCFMask( inst )

source

RvSpectML.CCF.TopHatCCFMask — Method.

Functor for returning 1 for any Δv <= width.

source

Instrument specific

RvSpectML.NEID.NEID1D — Type.

Trait for 1D Extracted spectra from NEID

source

RvSpectML.NEID.NEID2D — Type.

Trait for 2D Extracted spectra from NEID

source

RvSpectML.TheoreticalInstrument.TheoreticalInstrument1D — Type.

Trait for 1D spectra from a theoretical instrument

source

RvSpectML.TheoreticalInstrument.TheoreticalInstrument2D — Type.

Trait for a 2D spectra from a theoretical instrument

source