Functions

Functions Exported by RvSpectML

Functions Exported by RvSpectML

General purpose

RV-Related Algorithms

Interpolation Algorithms

Instrument-specific

Other

General purpose

Base.length — Method.

Return number of ChunkLists in a ChunkListTimeseries

source

Base.length — Method.

Return number of chunks in ChunkList

source

RvSpectML.absorption_line — Method.

absorption_line(x; mid, width, depth)

Return a Gaussian absorption line profile evaluated at x.

source

RvSpectML.apply_doppler_boost! — Function.

apply_doppler_boost!(spectrum, doppler_factor) -> typeof(spectrum)
apply_doppler_boost!(spectra, df) -> typeof(spectra)

Apply Doppler boost to spectra's λ's and update its metadata[:doppler_factor], so it will know how to undo the transform.

Arguments:

spectrum::AbstractSpectra: spectrum to be boosted

doppler_factor::Real: boost factor (1 = noop)

or

spectra::AbstractArray{<:AbstractSpectra}: spectra to be boosted

df::DataFrame: provides :drift and :ssb_rv (in m/s) for calculating the Doppler boost for each spectrum

TODO: Improve documentation formatting. This can serve as a template.

source

RvSpectML.bin_spectra_consecutive — Method.

binspectraconsecutive Bins consecutive spectra from a SpectralTimeSeriesCommonWavelengths object

WARNING: Simply takes consecutive spectra, so some bins may be from spectra that weren't taken close together. TODO: Create version that pays attention to timestamps.

source

RvSpectML.bin_spectra_max_Δt — Method.

binspectramax_Δt Bins spectra from a SpectralTimeSeriesCommonWavelengths object with a maximum spacing between observation times

source

RvSpectML.bin_spectra_nightly — Method.

binspectranightly Bins spectra from a SpectralTimeSeriesCommonWavelengths object with a maximum spacing between observation times

source

RvSpectML.bin_times_consecutive — Method.

bintimesconsecutive( times, n ) Computes mean times from conseuctive bins of n times (to go with binconsecutivespectra). Returns floor(length(times)/n) elements.

WARNING: Simply takes consecutive times, so some bins may be from spectra that weren't taken close together. TODO: Create version that pays attention to timestamps.

source

RvSpectML.calc_d2fluxdlnlambda2 — Method.

Estimate numerical second derivative of fluxes given wavelengths.

source

RvSpectML.calc_dfluxdlnlambda — Method.

Estimate numerical derivative of fluxes given wavelengths.

source

RvSpectML.calc_doppler_factor — Function.

calcdopplerfactor(rv; v_perp)

Return the Doppler boost factor (non-relativistic) for rv in km/s.

source

RvSpectML.calc_mean_dfluxdlnlambda — Method.

Return mean numerical derivative of fluxes based on a common set of wavelengths. Inputs: flux & var (2d) and λ (1d)

source

RvSpectML.calc_mean_spectrum — Method.

Return mean flux (averaging over observations at different times, variance weighted) based on a common set of wavelengths. Inputs: flux & var (2d: pixel, time)

source

RvSpectML.calc_normalization — Method.

Calc normalization of spectra based on average flux in a ChunkList.

source

RvSpectML.code_to_include_param_jl — Function.

codetoreadparamjl( pathto_search )

Returns a Code object. After res = code_toread_param_jl( path_to_search ), execute eval(res) to actually include the param.jl file. This is useful since it allows variables to be placed into caller's namespace.

Warning: Malicious users could insert arbitrary code into param.jl. Don't be a malicous user.

source

RvSpectML.find_overlapping_chunks — Method.

Return indices of any chunks in df that have overlapping lambdahi[i] and lambdalo[i+1].

source

RvSpectML.get_λ_range — Method.

Return the largest minimum wavelength and smallest maximum wavelength across an array of spectra. Calls getλrange(AbstractSpectra2D) that should be specialized for each instrument.

source

RvSpectML.make_chunk_list — Method.

Return a ChunkList of best regions of spectrum with lines in lineline. linelist is a DataFrame containing :lambdalo and :lambdahi. Pads edges by Δ.

source

RvSpectML.make_clean_line_list_from_tellurics_expres — Method.

makecleanlinelistfromtelluricsexpres(linelist, expresdata; Δvtoavoid_tellurics ) Returns a new line list that excludes lines with telluric contamination. Inputs:

line_list: Dataframe containing field lambda

expres_data: Array of EXPRES spectra

Δvtoavoid_tellurics: in m/s

Outputs:

linelistwithouttellurics: DataFrame with fields, lambda, weight, lambdalo, and lambda_hi.

Warning: Assumes a tellurics value in metadata for each spectra, such as is provided by EXPRES.

source

RvSpectML.make_grid_for_chunk — Method.

makegridfor_chunk Create a range with equal spacing between points with end points set based on union of all chunks in timeseries.

Arguments:

timeseries: ChunkListTimeseries

chunk index:

oversample_factor: (1)

source

RvSpectML.make_manifest — Method.

makemanifest(targetsubdir::String, Inst::Module; [opts] ) Returns a dataframe containing a list of files to be read and some metadata (e.g., observation times)

Optional arguements

maxspectratouse (defaultmaxspectrato_use)

pathstosearch (defaultpathsto_search)

verbose = true

Warning: Malicious users could insert arbitrary code into data_paths.jl. Don't be a malicous user.

source

RvSpectML.make_orders_into_chunks — Function.

makeordersinto_chunks

Return a ChunkList with a region of spectrum from each order in orderstouse.

Arguments

spectra<:AbstractSpectra

inst: Instrument trait that provides default values

Optional arguments

orderstouse: Range or Array (orderstouse(inst))

pixelstouse: Array of Ranges (each from mincol to maxcol)

or

mincol: (mincol_default(inst,order)) and

maxcol: (maxcol_default(inst,order))

source

RvSpectML.make_spectral_time_series_common_wavelengths_with_selected_times — Method.

makespectraltimeseriescommonwavelengthswithselectedtimes( input, timeidx ) Returns a SpectralTimeSeriesCommonWavelengths, retaining only those times and spectra specified by timeidx.

source

RvSpectML.make_template_spectra — Method.

maketemplatespectra( chunklisttimeseries, [ options ] ) Combine portions of spectra in a ChunkListTimeseries into a template stored as a SpectralTimeSeriesCommonWavelengths. Optionally remove radial velocities before combining spectra based on the :rv_est field in metadata.

Inputs:

chunklisttimeseries

Optional Parametes:

removervest (true)

oversample_factor (1)

smooth_factor (1)

alg (:TemporalGP)

Outputs

matrix: timeseries interpolated to each wavelength in the form of a SpectralTimeSeriesCommonWavelengths

f_mean: mean flux

var_mean: variance of mean flux (WARNING: Not computed accurately yet)

deriv: dflux/dlnλ averaged over spectra

deriv2: d²flux/dlnλ² averaged over spectra

source

RvSpectML.merge_chunks — Method.

Return DataFrame with specifications for each chunk which will contain one or more lines. Input: linelist a DataFrame with: - lambdalo, lambdahi, lambda, depth Output: DataFrame with - lambdalo & lambdahi: boundaries for chunk - lambda & linedepths: arrays with info about each line

source

RvSpectML.normalize_spectrum! — Method.

Normalize spectrum, multiplying fluxes by scale_fac.

source

RvSpectML.num_chunks — Method.

Number of chunks in first chunklist in chunklist_timeseries.

source

RvSpectML.plot_spectrum_chunks — Method.

plotspectrumchunks( chunklist, chunks ) Returns a plot based of flux versus wavelength for selected chunks from the spectra in chunklist.

source

RvSpectML.read_header — Method.

Read header from FITS file and return Dict with contents.

source

RvSpectML.read_linelist_espresso — Method.

Read line list in ESPRESSO csv format.

ESPRESSO format: lambda and weight. Warning: ESPRESSO masks don't provide line depth and sometimes include one entry for a blend of lines.

source

RvSpectML.read_linelist_vald — Method.

Read line list in VALD csv format. VALD format: lambdalo, lambdaahi and depth.

source

RvSpectML.read_manifest — Method.

Read manifest containing filename, bjd, target, and optionally additional metadata from CSV file.

source

RvSpectML.read_mask_espresso — Method.

Read mask in ESPRESSO csv format.

ESPRESSO format: lambda and weight. Warning: ESPRESSO masks don't provide line depth and sometimes include one entry for a blend of lines.

source

RvSpectML.read_mask_vald — Method.

Read mask in VALD csv format. VALD format: lambdalo, lambdaahi and depth.

source

RvSpectML.rms_rvs_within_night — Method.

rmsrvwithin_night(times, rvs) Return RMS of RVs taken within the same night

source

RvSpectML.λ_air_to_vac — Method.

Convert air wavelength (in Å) to vacuum wavelength Ref: https://www.astro.uu.se/valdwiki/Air-to-vacuum%20conversion VALD3 tools use the following solution derived by N. Piskunov

source

RvSpectML.λ_vac_to_air — Method.

Convert vacuum wavelength (in Å) to air wavelength Ref: Donald Morton (2000, ApJ. Suppl., 130, 403) via https://www.astro.uu.se/valdwiki/Air-to-vacuum%20conversion

source

RV-Related Algorithms

RvSpectML.CCF.calc_ccf_chunk — Method.

calcccfchunk( chunk, ccf_plan ) Convenience function to compute CCF for one chunk of spectrum.

Inputs:

chunk

Optional Arguments:

ccf_plan (BasicCCFPlan())

Return:

CCF for one chunk of spectrum, evaluated using mask_shape and line list from ccf plan

source

RvSpectML.CCF.calc_ccf_chunklist — Method.

calcccfchunklist ( chunklist, ccf_plans ) Convenience function to compute CCF based on a spectrum's chunklist.

Inputs:

chunklist

vector of ccf plans (one for each chunk)

Optional Arguments:

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using the line list and mask_shape from the ccf plan for each chunk.

source

RvSpectML.CCF.calc_ccf_chunklist_timeseries — Method.

calcccfchunklisttimeseries( chunklisttimeseries, line_list ) Convenience function to compute CCF for a timeseries of spectra, each with a chunklist. Uses multiple threads if avaliable.

Inputs:

chunklist_timeseries

Optional Arguments:

ccf_plan (BasicCCFPlan())

verbose (false)

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using the ccfplan. Note that the ccfplan provided is used as a template for creating a custom ccfplan for each chunk that only includes lines that reliably appear in that order for all spectra in the chunklisttimeseries.

source

RvSpectML.CCF.calc_ccf_v_grid — Method.

calcccfvgrid( plan ) Return range with 2n+1 points between -vmax and v_max where CCF is to be evaluated. Units based on those in plan.

source

RvSpectML.CCF.ccf_1D! — Method.

ccf_1D!(ccf_out, λs, fluxes; ccf_plan )

Compute the cross correlation function of a spectrum with a mask.

Inputs:

ccfout: 1-d array of size length(calcccfvgrid(plan)) to store output

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

Optional Arguments:

plan: parameters for computing ccf (BasicCCFPlan()), including maskshape and linelist

plan.line_list: Each mask entry consists of the left and right end of tophats and a weight.

plan.mask_shape: shape of mask to use (currently only works with TopHatCCFMask())

source

RvSpectML.CCF.ccf_1D — Method.

ccf_1D( λs, fluxes; ccf_plan )

Compute the cross correlation function of a spectrum with a mask.

Inputs:

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

Optional Arguments:

ccf_plan: parameters for computing ccf (BasicCCFPlan())

Returns:

1-d array of size length(calcccfv_grid(plan))

source

RvSpectML.CCF.ccf_1D_expr! — Method.

ccf_1D_expr!(ccf_out, λs, fluxes; ccf_plan )

Compute the cross correlation function of a spectrum with a mask. Experimental version that should work with different mask shapes. Need to understand why difference before merging this in.

Inputs:

ccfout: 1-d array of size length(calcccfvgrid(plan)) to store output

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

Optional Arguments:

plan: parameters for computing ccf (BasicCCFPlan())

source

RvSpectML.CCF.ccf_1D_expr — Method.

ccf_1D_expr(ccf_out, λs, fluxes; ccf_plan )

Compute the cross correlation function of a spectrum with a mask. Experimental version to allow for flexible mask shapes.

Inputs:

λs: 1-d array of wavelengths

fluxes: 1-d array of fluxes

Optional Arguments:

ccf_plan: parameters for computing ccf (BasicCCFPlan())

Returns:

1-d array of size length(calcccfv_grid(plan))

source

RvSpectML.RVFromCCF.measure_rv_from_ccf — Method.

measure_rv_from_ccf(vels, ccf; fit_type="gaussian")

Fit a gaussian to the CCF to calculate a velocity using the specified method. Valid arguments for fit_type include "gaussian", "quadratic", and "centroid".

source

RvSpectML.RVFromCCF.measure_rvs_from_ccf — Method.

measure_rvs_from_ccf(vels, ccf; fit_type="gaussian")

At each time, compute the cross correlation function and fit to calculate a velocity using the specified method. Valid arguments for fit_type include "gaussian", "quadratic", and "centroid".

source

Interpolation Algorithms

RvSpectML.LinearInterpolation.make_interpolator_linear_var — Method.

Return interpolator for variances in spectra.

source

RvSpectML.SincInterpolation.spectra_interpolate — Method.

Original author: Joe Ninan Converted to Julia and optimized by Christian Gilbertson Further adapted/optimized by Eric Ford Additional optimizations possible by preallocating arrasy for minargs, Nminargs, minvalues, FilterValues and OldYCoords.

source

Instrument-specific

RvSpectML.EXPRES.read_data — Function.

Read EXPRES data from FITS file, and return in a Spectra2DBasic object.

source

RvSpectML.NEID.make_manifest — Method.

Create Dataframe containing filenames and key data for all files neid*.fits in directory

source

RvSpectML.NEID.read_barycentric_corrections! — Function.

Read CSV of NEID barycentric corrections, interpolate to bjd's in df and insert into df[:,ssb_rv].

source

RvSpectML.NEID.read_data — Function.

Read NEID (non-solar) data from FITS file, and return in a Spectra2DBasic object.

source

RvSpectML.NEID.read_drift_corrections! — Function.

Read CSV of NEID drift corrections, interpolate to bjd's in df and insert into df[:,drift].

source

RvSpectML.NEID.read_metadata — Method.

Create Dict containing filename and default metadata from file.

source

RvSpectML.NEID.read_solar_data — Function.

Read NEID Solar data from FITS file, and return in a Spectra2DBasic object.

source

Other