Internals

RvSpectML

Internals Types & Functions

RvSpectML

Modules

Exported Functions

Exported Types

RvSpectML

Internals Types & Functions

General purpose

Radial Velocity Related

Interpolation

Instrument specific

Index

General purpose

RvSpectML.ChunkWidthFuncOfλ — Type.

Functor to return a scale factor for chunk widths that depends on the chunk's central wavelength.

TODO: Replace with something useful, e.g., uses actual PSF behavior as function of wavelength

source

Base.getindex — Method.

Allow direct access to data, an AbstractArray of ChunkOfSpectrum's via [] operator

source

Base.setindex! — Method.

Allow direct access to data, an AbstractArray of ChunkOfSpectrum's via [] operator

source

RvSpectML.allequal — Method.

Return true if all elements of array are equal to each other.

source

RvSpectML.bin_times — Method.

bintimes() Bins times from a SpectralTimeSeriesCommonWavelengths object using the groupings from timeidx in the metadata.

source

RvSpectML.calc_d2fluxdlnlambda2! — Method.

Estimate numerical second derivative of fluxes given wavelengths.

source

RvSpectML.calc_dfluxdlnlambda! — Method.

Estimate numerical derivative of fluxes given wavelengths.

source

RvSpectML.calc_line_bisector_at_frac_depth — Method.

calclinebisectoratfracdepth(λ, flux, fracdepth ) Returns the line average of wavelengths (units of λ) at specified fractional line depth. Assumes continuum is the maximum flux provided.

source

RvSpectML.calc_line_width — Method.

calclinebisectoratfracdepth(λ, flux, absdepth ) Returns the line average of wavelengths (units of λ) at specified absolute line depth. Assumes continuum is the maximum flux provided.

source

RvSpectML.calc_line_width — Method.

calclinewidth(λ, flux; frac_depth ) Returns the line width (units of λ) for specified fractional line depth (default of 0.5). Assumes continuum is the maximum flux provided.

source

RvSpectML.calc_snr — Method.

Calculate total SNR in (region of) spectra.

source

RvSpectML.check_if_line_match — Method.

checkifline_match ( λ, list ; threshold ) Return true if list contains a wavelength differing from λ by no more than threshold (in units of Δλ/λ)

source

RvSpectML.find_cols_to_fit — Method.

Return a range of columns indices with wavelengths between linelo and linehi

source

RvSpectML.find_cols_to_fit — Method.

Return a range of columns indices with wavelengths within Δ of line_center

source

RvSpectML.find_line_best — Function.

Return (pixels, order) pair that contain "best" region of spectra, based on highest SNR.

source

RvSpectML.find_orders_with_line — Method.

Return list of all orders that contain a pixel with wavelength lambda

source

RvSpectML.find_orders_with_line — Method.

Return list of all orders that contain all pixels with wavelengths between goallo and goalhi

source

RvSpectML.find_pixels_for_line_in_chunk — Method.

Find pixels included in a range of wavelengths

source

RvSpectML.find_which_line_fits_in_line_list — Method.

findwhichlinefitsinlinelist( fitlist, linelist; threshold ) Return list of Bools indicatin which line(s) from fitlist match a line in linelist to within threshold (in units of Δλ/λ) Warning: Untested

source

RvSpectML.find_worst_telluric_in_each_chunk — Method.

findworsttelluricineachchunk( chunklisttimeseries, arrayof_spectra ) Returns a vector of the worst telluric (i.e., smallest value) within each chunk at any observations.

Warning: Assumes a tellurics value in metadata for each spectra, such as is provided by EXPRES.

source

RvSpectML.findall_line — Function.

Return list of (pixels, order) pairs that contain pixels with desireed wavelengths. Excludes locations that contain any pixels with var == NaN.

source

RvSpectML.findargminmax — Method.

findargminmax(a) Return (argmin, min, argmax, max) Adapapted from https://github.com/JuliaLang/julia/blob/697e782ab86bfcdd7fd15550241fe162c51d9f98/base/array.jl#L2191

source

RvSpectML.get_λs — Method.

" get_λs( grids::Vector{AbstractRange}, r::AbstractUnitRange ) Return range of wavelengths corresponding to indices in r from a vector of ranges. Useful for extracting wavelengths from a chunklist.

source

RvSpectML.interp_chunk_to_shifted_grid_linear! — Method.

interpchunktogridlinear!( fluxout, varout, chunkofspectrum, wavelengths ) Return spectra interpolated onto a grid of points using linear interpolation.

Arguments:

flux_out::AbstractArray

var_out::AbstractArray

chunk::AbstractChunkOfSpectrum

grid::AbstractRange or AbstractArray

Alters

flux_out

var_out::AbstractArray

Returns

flux_out

source

RvSpectML.interp_chunk_to_shifted_grid_sinc! — Method.

interpchunktogridsinc!( fluxout, varout, chunkofspectrum, wavelengths ) Return spectra interpolated onto a grid of points using sinc interpolation.

Arguments:

flux_out::AbstractArray

var_out::AbstractArray

chunk::AbstractChunkOfSpectrum

grid::AbstractRange or AbstractArray

Alters

flux_out

var_out::AbstractArray

Returns

flux_out

source

RvSpectML.interp_linear — Method.

interp_linear(;x1::T1,x2::T1,y1::T2,y2::T2,xpred::T1) Return result of simple linear interpolant at xpred. Does not test that xpred is between x1 and x2.

source

RvSpectML.make_template_spectra — Method.

maketemplatespectra( chunklisttimeseries, [ options ] ) Combine portions of spectra in a ChunkListTimeseries into a template stored as a SpectralTimeSeriesCommonWavelengths. Optionally remove radial velocities before combining spectra based on the :rv_est field in metadata.

Inputs:

chunklisttimeseries

Optional Parametes:

removervest (true)

oversample_factor (1)

smooth_factor (1)

alg (:TemporalGP)

Outputs

matrix: timeseries interpolated to each wavelength in the form of a SpectralTimeSeriesCommonWavelengths

f_mean: mean flux

var_mean: variance of mean flux (WARNING: Not computed accurately yet)

deriv: dflux/dlnλ averaged over spectra

deriv2: d²flux/dlnλ² averaged over spectra

source

RvSpectML.multiple_append! — Method.

A generalized version of the built in append!() function By Christian Gilbertson?

TODO: Ask Christian what the purpose of this is relative to std append

source

RvSpectML.normalize_spectra! — Method.

Normalize each spectrum based on sum of fluxes in chunk_timeseries region of each spectrum.

source

RvSpectML.predict_intrinsic_stellar_line_width — Method.

Estimate line width based on stellar Teff (K) and optionally v_rot (km/s). Output in km/s.

source

RvSpectML.read_metradata_from_fits — Function.

Read metadata in FITS header and return data for keys in fields_str/fields as a Dict.

source

RvSpectML.searchsortednearest — Function.

searchsortednearest(a<:AbstractVector, x::Real; assumesorted = false ) searchsortednearest(a<:AbstractVector, x<:AbstractVector; assumesorted = false )

Find the index of vector a where the value of a is closest to x. All vectors are assumed to already be sorted. To turn off assertions, set assume_sorted to true.

Credit: traktofon @ https://discourse.julialang.org/t/findnearest-function/4143/4 Vector Vector version by Christian Gilbertson? issorted assertion and optional assume_sorted added by Eric Ford

source

Radial Velocity Related

RvSpectML.CCF.CosCCFMask — Type.

CosCCFMask Cosine mask with one parameter, it's quarter period, i.e., where to truncate it, as a velocity in m/s. Mask weights are stored separately in a line list.

source

RvSpectML.CCF.CosCCFMask — Method.

CosCCFMask( inst )

source

RvSpectML.CCF.CosCCFMask — Method.

Functor for returning PSF for Δv <= half_width.

source

RvSpectML.CCF.GaussianCCFMask — Type.

GaussianCCFMask A truncated Gaussian mask with two parameters, it's stdandard deviation and where to truncate it, both as a velocity in m/s. Mask weights are stored separately in a line list.

source

RvSpectML.CCF.GaussianCCFMask — Method.

GaussianCCFMask( inst ; scale_factor )

source

RvSpectML.CCF.GaussianCCFMask — Method.

Functor for returning PSF for Δv <= half_width.

source

Base.length — Method.

Return length of line linst.

source

RvSpectML.CCF.calc_ccf_chunk_expr — Method.

calcccfchunkexpr( chunk, ccfplan ) Convenience function to compute CCF for one chunk of spectrum. Uses experimental version of projectmaskexpr!. Need to understand why difference before merging this in.

Inputs:

chunk

Optional Arguments:

ccf_plan (BasicCCFPlan())

Return:

CCF for one chunk of spectrum, evaluated using mask_shape and line list from ccf plan

source

RvSpectML.CCF.calc_ccf_chunklist_expr — Method.

calcccfchunklistexpr ( chunklist, ccfplans ) Convenience function to compute CCF based on a spectrum's chunklist. Experimental version, testing for general mask shapes.

Inputs:

chunklist

vector of ccf plans (one for each chunk)

Optional Arguments:

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using the line list and mask_shape from the ccf plan for each chunk.

source

RvSpectML.CCF.calc_ccf_chunklist_timeseries_expr — Method.

calcccfchunklisttimeseriesexpr( chunklisttimeseries, linelist ) Convenience function to compute CCF for a timeseries of spectra, each with a chunklist. Uses multiple threads if avaliable. Experimental version trying to generalize mask shape

Inputs:

chunklist_timeseries

Optional Arguments:

ccf_plan (BasicCCFPlan())

verbose (false)

Return:

CCF summed over all chunks in a spectrum's chunklist, evaluated using the ccfplan. Note that the ccfplan provided is used as a template for creating a custom ccfplan for each chunk that only includes lines that reliably appear in that order for all spectra in the chunklisttimeseries.

source

RvSpectML.CCF.calc_doppler_factor — Method.

calc_doppler_factor(vel)

Compute the longitudinal relativistic doppler factor given a velocity in meters per second.

source

RvSpectML.CCF.calc_order_ccf_chunklist_timeseries — Method.

calcorderccfchunklisttimeseries( chunklisttimeseries, ccfplan ) Convenience function to compute separate CCFs for each chunk (typically an order) of each spectrum in a timeseries. CCF is evaluated using line list and mask_shape provided by the ccf plan for each chunk. Uses multiple threads if avaliable.

Inputs:

chunklist_timeseries

Optional Arguments:

ccf_plan (BasicCCFPlan())

Return:

A 3-d array containing the CCF at each (velocity, chunk, spectrum) Note that the ccfplan provided is used as a template for creating a custom ccfplan for each chunk that only includes lines that reliably appear in that order for all spectra in the chunklist_timeseries.

source

RvSpectML.CCF.calc_order_ccfs_chunklist — Method.

calcorderccfschunklist ( chunklisttimeseries, linelist ) Convenience function to compute separate CCFs for each chunk (typically an order) in a spectrum. CCF is evaluated using line list and maskshape provided by the ccf plan for each chunk.

Inputs:

chunklist_timeseries

vector of ccf plans (one for each chunk)

Return:

A 2-d array containing the CCF at each (velocity, chunk)

source

RvSpectML.CCF.find_bin_edges_opt — Method.

findbinedges( pixel_centers )

Internal function used by project_mask!. TODO: OPT may be able to eliminate need for this memory allocation.

source

RvSpectML.CCF.project_mask_expr! — Method.

projectmaskexpr!( output, λs, ccfplan; shiftfactor, workspace )

Compute the projection of the mask onto the 1D array of wavelengths (λs) at a given shift factor (default: 1). The mask is computed from the ccfplan, including a linelist and mask_shape (default: tophat).

Note: Preliminary implementation for other maskshapes is in projectmask_expr!

source

RvSpectML.CCF.project_mask_opt! — Method.

project_mask_opt!( output, λs, ccf_plan; shift_factor, workspace )

Compute the projection of the mask onto the 1D array of wavelengths (λs) at a given shift factor (default: 1). The mask is computed from the ccfplan, including a linelist and mask_shape (default: tophat).

Note: Preliminary implementation for other maskshapes is in projectmask_expr!

source

RvSpectML.RVFromCCF.measure_rv — Method.

measure_rv(wavs, spec, mask; fit_type="gaussian")

TODO: Need to update for CCF refactoring.

Compute the cross correlation function and fit to calculate a velocity using the specified method. Valid arguments for fit_type include "gaussian", "quadratic", and "centroid".

source

RvSpectML.RVFromCCF.prelim_fit — Method.

source

RvSpectML.RVFromCCF.rv_from_ccf_bestfit — Method.

Report RV based on minimum of CCF WARNING: This is highly pixelated and should be used for special purposes (e.g., initial guesses).

source

RvSpectML.RVFromCCF.rv_from_ccf_centroid — Method.

Report RV based on centroid near minimum of CCF. WARNING: Something's fishy with these.

source

RvSpectML.RVFromCCF.rv_from_ccf_gaussian — Method.

Report RV base don fit Gaussian near minimum of CCF.

source

RvSpectML.RVFromCCF.rv_from_ccf_quadratic — Method.

Report RV based on fitting quadratic near minimum of CCF.

source

Interpolation

RvSpectML.SincInterpolation.create_filter_curve_orig — Method.

Original author: Joe Ninan Converted to Julia and optimized by Christian Gilbertson & Eric Ford Returns a cubit interpolator for windowed sinc Filter curve. noofpoints: number of intepolation points to use in cubic inteprolator

source

Instrument specific

RvSpectML.EXPRES.make_manifest — Method.

Create Dataframe containing filenames and key data for all files neid*.fits in directory

source

RvSpectML.EXPRES.read_metadata — Method.

Create Dict containing filename and default metadata from file.

source

RvSpectML.NEID.read_differential_extinctions! — Function.

Read space delimited file with differential extinction corrections, interpolate to bjd's in df and insert into df[:,diffextrv].

source